+++ cmd/generate/action/runner.go
@@ -14,6 +14,7 @@ import (
 	"github.com/giantswarm/micrologger"
 	"github.com/spf13/cobra"
 
+	actiontemplate "github.com/giantswarm/awscnfm/cmd/generate/action/template"
 	templatecmd "github.com/giantswarm/awscnfm/cmd/generate/action/template/cmd"
 	templatecmdexecute "github.com/giantswarm/awscnfm/cmd/generate/action/template/cmd/execute"
 	templatecmdexplain "github.com/giantswarm/awscnfm/cmd/generate/action/template/cmd/explain"
@@ -47,45 +48,65 @@ func (r *runner) Run(cmd *cobra.Command, args []string) error {
 }
 
 func (r *runner) run(ctx context.Context, cmd *cobra.Command, args []string) error {
-	data := struct {
-		Action  string
-		Cluster string
-	}{
+	data := actiontemplate.Data{
 		Action:  r.flag.Action,
 		Cluster: r.flag.Cluster,
 	}
 
-	// templates is a predefined list of lists for debugging reasons. When
-	// defining map[string]string for the key-value pairs the order of items
-	// changes since go maps are not deterministic.
-	templates := [][]string{
-		{templatecmdexecute.CommandBase, templatecmdexecute.CommandContent},
-		{templatecmdexecute.ErrorBase, templatecmdexecute.ErrorContent},
-		{templatecmdexecute.FlagBase, templatecmdexecute.FlagContent},
-		{templatecmdexecute.RunnerBase, templatecmdexecute.RunnerContent},
+	{
+		// templates is a predefined list of lists for debugging reasons. When
+		// defining map[string]string for the key-value pairs the order of items
+		// changes since go maps are not deterministic.
+		templates := [][]string{
+			{templatecmdexecute.CommandBase, templatecmdexecute.CommandContent},
+			{templatecmdexecute.ErrorBase, templatecmdexecute.ErrorContent},
+			{templatecmdexecute.FlagBase, templatecmdexecute.FlagContent},
+			{templatecmdexecute.RunnerBase, templatecmdexecute.RunnerContent},
 
-		{templatecmdexplain.CommandBase, templatecmdexplain.CommandContent},
-		{templatecmdexplain.ErrorBase, templatecmdexplain.ErrorContent},
-		{templatecmdexplain.FlagBase, templatecmdexplain.FlagContent},
-		{templatecmdexplain.RunnerBase, templatecmdexplain.RunnerContent},
+			{templatecmdexplain.CommandBase, templatecmdexplain.CommandContent},
+			{templatecmdexplain.ErrorBase, templatecmdexplain.ErrorContent},
+			{templatecmdexplain.FlagBase, templatecmdexplain.FlagContent},
+			{templatecmdexplain.RunnerBase, templatecmdexplain.RunnerContent},
 
-		{templatecmd.CommandBase, templatecmd.CommandContent},
-		{templatecmd.ErrorBase, templatecmd.ErrorContent},
-		{templatecmd.FlagBase, templatecmd.FlagContent},
-		{templatecmd.RunnerBase, templatecmd.RunnerContent},
+			{templatecmd.CommandBase, templatecmd.CommandContent},
+			{templatecmd.ErrorBase, templatecmd.ErrorContent},
+			{templatecmd.FlagBase, templatecmd.FlagContent},
+			{templatecmd.RunnerBase, templatecmd.RunnerContent},
+		}
 
-		{templatepkgaction.ErrorBase, templatepkgaction.ErrorContent},
-		{templatepkgaction.ExecutorBase, templatepkgaction.ExecutorContent},
-		{templatepkgaction.ExecutorCustomBase, templatepkgaction.ExecutorCustomContent},
-		{templatepkgaction.ExplainerBase, templatepkgaction.ExplainerContent},
-		{templatepkgaction.ExplainerCustomBase, templatepkgaction.ExplainerCustomContent},
+		err := write(data, templates, "cmd/%s/%s")
+		if err != nil {
+			return microerror.Mask(err)
+		}
 	}
 
+	{
+		// templates is a predefined list of lists for debugging reasons. When
+		// defining map[string]string for the key-value pairs the order of items
+		// changes since go maps are not deterministic.
+		templates := [][]string{
+			{templatepkgaction.ErrorBase, templatepkgaction.ErrorContent},
+			{templatepkgaction.ExecutorBase, templatepkgaction.ExecutorContent},
+			{templatepkgaction.ExecutorCustomBase, templatepkgaction.ExecutorCustomContent},
+			{templatepkgaction.ExplainerBase, templatepkgaction.ExplainerContent},
+			{templatepkgaction.ExplainerCustomBase, templatepkgaction.ExplainerCustomContent},
+		}
+
+		err := write(data, templates, "pkg/action/%s/%s")
+		if err != nil {
+			return microerror.Mask(err)
+		}
+	}
+
+	return nil
+}
+
+func write(data actiontemplate.Data, templates [][]string, dirfmt string) error {
 	for _, l := range templates {
 		base := l[0]
 		cont := l[1]
 
-		path, err := filepath.Abs(filepath.Join(fmt.Sprintf("cmd/%s/%s", data.Cluster, data.Action), base))
+		path, err := filepath.Abs(filepath.Join(fmt.Sprintf(dirfmt, data.Cluster, data.Action), base))
 		if err != nil {
 			return microerror.Mask(err)
 		}
@@ -109,7 +130,7 @@ func (r *runner) run(ctx context.Context, cmd *cobra.Command, args []string) err
 		var write bool
 		{
 			var exists bool
-			_, err = os.Stat(name)
+			_, err = os.Stat(path)
 			if os.IsNotExist(err) {
 				// fall through, exists is already false
 			} else if err != nil {
