+++ cmd/generate/action/template/data.go
@@ -1,0 +1,6 @@
+package template
+
+type Data struct {
+	Action  string
+	Cluster string
+}
